+++ R/derive_locf_records.R
@@ -17,8 +17,14 @@
 #'
 #'   For each group defined by `by_vars` those observations from `dataset_expected_obs`
 #'   are added to the output dataset which do not have a corresponding observation
-#'   in the input dataset or for which `AVAL` is NA for the corresponding observation
+#'   in the input dataset or for which `analysis_var` is NA for the corresponding observation
 #'   in the input dataset.
+#'
+#' @param analysis_var Analysis variable.
+#'
+#'   *Default*: `AVAL`
+#'
+#'   *Permitted Values*: a variable
 #'
 #' @param order List of variables for sorting a dataset
 #'
@@ -27,9 +33,9 @@
 #'
 #' @param keep_vars Variables that need carrying the last observation forward
 #'
-#'   Keep variables that need carrying the last observation forward other than `AVAL`
+#'   Keep variables that need carrying the last observation forward other than `analysis_var`
 #'   (e.g., `PARAMN`, `VISITNUM`). If by default `NULL`, only variables specified in
-#'   `by_vars` and `AVAL` will be populated in the newly created records.
+#'   `by_vars` and `analysis_var` will be populated in the newly created records.
 #'
 #' @author G Gayatri
 #'
@@ -37,9 +43,9 @@
 #' by_vars parameter) those observations from dataset_expected_obs are added to
 #' the output dataset
 #' - which do not have a corresponding observation in the input dataset or
-#' - for which `AVAL` is NA for the corresponding observation in the input dataset.
+#' - for which `analysis_var` is NA for the corresponding observation in the input dataset.
 #'
-#'   For the new observations, `AVAL` is set to the non-missing `AVAL` of the
+#'   For the new observations, `analysis_var` is set to the non-missing `analysis_var` of the
 #'   previous observation in the input dataset (when sorted by `order`) and
 #'   `DTYPE` is set to "LOCF".
 #'
@@ -105,9 +111,11 @@
 derive_locf_records <- function(dataset,
                                 dataset_expected_obs,
                                 by_vars,
+                                analysis_var = AVAL,
                                 order,
                                 keep_vars = NULL) {
   #### Input Checking ####
+  analysis_var <- assert_symbol(enquo(analysis_var))
 
   # Check if input parameters is a valid list of variables
   assert_vars(by_vars, optional = TRUE)
@@ -118,7 +126,7 @@ derive_locf_records <- function(dataset,
   assert_data_frame(dataset_expected_obs)
   assert_data_frame(
     dataset,
-    required_vars = quo_c(by_vars, extract_vars(order), chr2vars(colnames(dataset_expected_obs)))
+    required_vars = quo_c(by_vars, analysis_var, extract_vars(order), chr2vars(colnames(dataset_expected_obs)))
   )
   assert_data_frame(
     dataset,
@@ -143,12 +151,12 @@ derive_locf_records <- function(dataset,
 
   ##### Add LOCF records ####
 
-  # Split input dataset into the missing and non-missing AVAL records
+  # Split input dataset into the missing and non-missing analysis_var (e.g., AVAL) records
   aval_missing <- dataset %>%
-    filter(is.na(AVAL))
+    filter(is.na(!!analysis_var))
 
   aval_not_missing <- dataset %>%
-    drop_na(AVAL)
+    drop_na(!!analysis_var)
 
 
   # Get the variable names from the expected observation dataset
@@ -157,7 +165,7 @@ derive_locf_records <- function(dataset,
 
 
   # Get unique combination of visits/timepoints per parameter per subject
-  # from the original input dataset (with non-missing AVAL)
+  # from the original input dataset (with non-missing analysis_var)
   advs_unique_original <- aval_not_missing %>%
     select(all_of(exp_obs_vars)) %>%
     distinct()
@@ -166,14 +174,14 @@ derive_locf_records <- function(dataset,
   tmp_dtype <- get_new_tmp_var(exp_obsv, prefix = "tmp_dtype")
 
   # Get all the expected observations that are to be added to the input
-  # dataset (with non-missing AVAL)
+  # dataset (with non-missing analysis_var)
   advs_exp_obsv3 <- exp_obsv %>%
     mutate(!!tmp_dtype := "LOCF") %>%
     anti_join(advs_unique_original, by = c(exp_obs_vars))
 
-  # Merge the expected observations with the input dataset (with non-missing AVAL)
+  # Merge the expected observations with the input dataset (with non-missing analysis_var)
   # Arrange the dataset by 'order' and group it by 'by_vars'
-  # Use fill() to fill the AVAL from the previous observation for the newly added records
+  # Use fill() to fill the analysis_var from the previous observation for the newly added records
 
 
   aval_not_missing_locf <- aval_not_missing %>%
@@ -190,11 +198,11 @@ derive_locf_records <- function(dataset,
   aval_not_missing_locf <- aval_not_missing_locf %>%
     arrange(!!!by_vars, !!!order) %>%
     group_by(!!!by_vars) %>%
-    fill("AVAL", !!!keep_vars) %>%
+    fill(!!analysis_var, !!!keep_vars) %>%
     ungroup()
 
 
 
-  # Output dataset - merge the AVAL missing with non-missing+newly added LOCF records
+  # Output dataset - merge the analysis_var missing with non-missing+newly added LOCF records
   bind_rows(aval_not_missing_locf, aval_missing)
 }