+++ tests/testthat/test-derive_locf_records.R
@@ -263,9 +263,9 @@ test_that("derive_locf_records Test 5: visit variables are parameter dependent",
 ## Test 6: populate VISITNUM for LOCF records ----
 test_that("derive_locf_records Test 6: populate VISITNUM for LOCF records", {
   input <- tibble::tribble(
-    ~STUDYID, ~USUBJID, ~PARAMCD, ~PARAM, ~AVAL, ~AVISITN, ~AVISIT, ~VISITNUM,
-    "TEST01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 79, 0, "BASELINE", 0,
-    "TEST01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 130, 0, "BASELINE", 0
+    ~STUDYID, ~USUBJID, ~PARAMCD, ~PARAM, ~AVALC, ~AVISITN, ~AVISIT, ~VISITNUM,
+    "TEST01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", "79", 0, "BASELINE", 0,
+    "TEST01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", "130", 0, "BASELINE", 0
   )
 
   advs_expected_obsv <- tibble::tribble(
@@ -280,9 +280,9 @@ test_that("derive_locf_records Test 6: populate VISITNUM for LOCF records", {
   expected_output <- bind_rows(
     input,
     tibble::tribble(
-      ~STUDYID, ~USUBJID, ~PARAMCD, ~PARAM, ~AVAL, ~AVISITN, ~AVISIT, ~VISITNUM,
-      "TEST01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 79, 2, "WEEK 2", 0,
-      "TEST01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 130, 2, "WEEK 2", 0
+      ~STUDYID, ~USUBJID, ~PARAMCD, ~PARAM, ~AVALC, ~AVISITN, ~AVISIT, ~VISITNUM,
+      "TEST01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", "79", 2, "WEEK 2", 0,
+      "TEST01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", "130", 2, "WEEK 2", 0
     ) %>%
       mutate(DTYPE = "LOCF")
   )
@@ -292,6 +292,7 @@ test_that("derive_locf_records Test 6: populate VISITNUM for LOCF records", {
     input,
     dataset_expected_obs = advs_expected_obsv,
     by_vars = vars(STUDYID, USUBJID, PARAM, PARAMCD),
+    analysis_var = AVALC,
     order = vars(AVISITN, AVISIT),
     keep_vars = vars(VISITNUM)
   )